Update qq.R
Hi Stephen,

The suggested code change aims to prevent the over plotting of data. This significantly speeds up the plotting and is good for rasterized output such as PDF and EPS which do not handle hundreds of thousands of data-points well.  The downside is a minor increase in the processing speed of the function due to the for loop.

Hope this can make it into any updated version.

All the best
Mike